+++ src/plone/app/mosaic/browser/static/js/mosaic.toolbar.js
@@ -527,8 +527,8 @@ define([
           setTimeout(function(){
             // we want to do this on a delay to prevent conflict errors when
             // editing a tile and this is called at the exact same time
-            $("#form-widgets-ILayoutAware-customLayout, " +
-              "[name='form.widgets.ILayoutAware.customLayout']").blur();
+            $("#form-widgets-ILayoutAware-customContentLayout, " +
+              "[name='form.widgets.ILayoutAware.customContentLayout']").blur();
           }, 1000);
           lastChange = (new Date()).getTime();
         }